Fixed issues with dynamic Gaussian blur shader.
The kernel size from the dynamic Gaussian blur asset has been changed from 11 to 5. An option to adjust this kernel will be availble int he future.

We also fixed a bug which was causing the shader to apply itself multiple times and shift the content of the screen. The asset should perform far better now, and no longer drag the image as it did before.